Bruno, nos quedamos libres

+++ src/main/java/com/bda/TrabajoPractico/repositorios/EstacionRepository.java
@@ -3,5 +3,5 @@ package com.bda.TrabajoPractico.repositorios;
 import com.bda.TrabajoPractico.Entidades.Estacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstacionRepository extends JpaRepository<Estacion,Integer> {
+public interface EstacionRepository extends JpaRepository<Estacion,Long> {
 }
